requirements.txt
pyyaml
requests
pypredict
flask
pysqlite3
gunicorn
python-dateutil
